+++ Fossil/src/Game.cpp
@@ -7,8 +7,9 @@ int main() {
 	logMessage(". . . Starting . . . \n", "gamelogs.txt", true);
 
 	//	Create an array to store the settings and get settings from gameprefs
-	std::vector<std::string> displaySettings = getSettings("Display");
-	std::vector<std::string> graphicsSettings = getSettings("Graphics");
+	std::vector<std::string> displaySettings = getSettings("../../../gameprefs.txt", "Display");
+	std::vector<std::string> graphicsSettings = getSettings("../../../gameprefs.txt", "Graphics");
+	std::vector<std::string> multSettings = getSettings("../../../multiplayerprefs.txt", "Multiplayer");
 
 	//	Log message noting that the game has exited
 	logMessage("\n . . . Ending . . .", "gamelogs.txt", false);